- Big Update
1. Add Editor project
    a. Add SceneHierarchy Panel and Propertis Panel
    b. Implement Docking space
2. Update Input module
    a. Implement KeyCode, MouseCode
    b. Improve Event system with new Input module
3. Add Entity-Component System
4. Add Native Scripting Module
5. Update Camera System
    a. Remove CameraController class
    b. Update Camera class so that it fits in ECS
    c. Update Camera class so that it support both 2D and 3D
6. Update Render System
    a. Add Batch rendering for sprites and textures
    b. Add SubTexture for 2D Renderer
    c. Add FBO
    d. Wrap all render code in Renderer2D class
    e. Make RenderCommand class only visible internally
    f. Remove Renderer class
7. Remove Redundant Libraries
    a. Remove IrrKlang
8. Fix Bugs
    a. Fix wrong case in Vertex Attribute Pointer setting
9. Fix Warnings
    a. Fix mismatch in dependencies' build options
    b. Fix type cast in Buffers.h
    c. Fix missing check for wrong vertex data type in VAO
10. Code Maintenance
    a. Add iterator wrapper for LayerStack
    b. Replace raw pointers with Scope<> and Ref<>
    c. Remove all redundant inline
10. Update Project scripts
    a. Remove code to do with IrrKlang
    b. Add deletion of imgui.ini

+++ Illusion/src/Engine/Render/Animation/SequenceAnimation.cpp
@@ -9,7 +9,7 @@ namespace Illusion
 		:m_FrameCount(framecount), m_CurrentFrame(0.0f), m_FolderPath(folderpath), m_Name(name), m_End(false)
 	{
 		char filepath[256];
-		for (int i = 0; i < m_FrameCount; i++)
+		for (uint32_t i = 0; i < m_FrameCount; i++)
 		{
 			memset(filepath, 0, 256);
 			sprintf(filepath, "%s/%s (%d).png", m_FolderPath.c_str(), m_Name.c_str(), i + 1);
@@ -35,7 +35,7 @@ namespace Illusion
 		{
 			if ((uint32_t)m_CurrentFrame == m_FrameCount)
 			{
-				m_CurrentFrame = m_FrameCount - 1;
+				m_CurrentFrame = (float)(m_FrameCount - 1);
 				m_End = true;
 			}
 		}
@@ -49,7 +49,7 @@ namespace Illusion
 		m_CurrentFrame += step;
 		if ((uint32_t)m_CurrentFrame == m_FrameCount)
 		{
-			m_CurrentFrame = m_FrameCount - 1;
+			m_CurrentFrame = (float)(m_FrameCount - 1);
 			m_End = true;
 		}
 	}
